+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [2.4.0] - 2023-06-30
+### Improvements
+- PHP 8.1 compatibility
+- Compatibility with Wordpress 6.2 and WooCommerce 7.7
+- Added a new setting that allows admin to input custom gateway URL.
+### Changed
+- Updated the SDK version to 1.7.0
+
 ## [2.3.2] - 2022-05-11
 ### Fixed
 - “Amount mismatch” error is triggered when a user purchases a product on recent versions of WooCommerce (WordPress 5.9.3 / Woocommerce 6.4.1)